Added option to register custom types for extensions method lookup Refactored ExpressionHelper
