+++ src/DotVVM.Framework.Tests.Common/Binding/StaticCommandCompilationTests.cs
@@ -61,7 +61,7 @@ namespace DotVVM.Framework.Tests.Binding
             var options = BindingParserOptions.Create<ValueBindingExpression>()
                 .AddImports(configuration.Markup.ImportedNamespaces);
 
-            var parser = new BindingExpressionBuilder(configuration.ServiceProvider.GetRequiredService<CompiledAssemblyCache>());
+            var parser = new BindingExpressionBuilder(configuration.ServiceProvider.GetRequiredService<CompiledAssemblyCache>(), configuration.ServiceProvider.GetRequiredService<MemberExpressionFactory>());
             var expressionTree = parser.ParseWithLambdaConversion(expression, context, options, expectedType);
             var jsExpression =
                 configuration.ServiceProvider.GetRequiredService<StaticCommandBindingCompiler>().CompileToJavascript(context, expressionTree);
