+++ src/DotVVM.Framework/Compilation/Binding/GeneralBindingPropertyResolvers.cs
@@ -28,13 +28,15 @@ namespace DotVVM.Framework.Compilation.Binding
         private readonly IBindingExpressionBuilder bindingParser;
         private readonly StaticCommandBindingCompiler staticCommandBindingCompiler;
         private readonly JavascriptTranslator javascriptTranslator;
+        private readonly MemberExpressionFactory memberExpressionFactory;
 
-        public BindingPropertyResolvers(IBindingExpressionBuilder bindingParser, StaticCommandBindingCompiler staticCommandBindingCompiler, JavascriptTranslator javascriptTranslator, DotvvmConfiguration configuration)
+        public BindingPropertyResolvers(IBindingExpressionBuilder bindingParser, StaticCommandBindingCompiler staticCommandBindingCompiler, JavascriptTranslator javascriptTranslator, DotvvmConfiguration configuration, MemberExpressionFactory memberExpressionFactory)
         {
             this.configuration = configuration;
             this.bindingParser = bindingParser;
             this.staticCommandBindingCompiler = staticCommandBindingCompiler;
             this.javascriptTranslator = javascriptTranslator;
+            this.memberExpressionFactory = memberExpressionFactory;
         }
 
         public ActionFiltersBindingProperty GetActionFilters(ParsedExpressionBindingProperty parsedExpression)
@@ -71,7 +73,7 @@ namespace DotVVM.Framework.Compilation.Binding
         {
             var valueParameter = Expression.Parameter(typeof(object), "value");
             var body = BindingCompiler.ReplaceParameters(binding.Expression, dataContext);
-            body = ExpressionHelper.UpdateMember(body, valueParameter);
+            body = memberExpressionFactory.UpdateMember(body, valueParameter);
             if (body == null)
             {
                 return null;
