+++ src/DotVVM.Framework/Compilation/Binding/MethodGroupExpression.cs
@@ -72,16 +72,16 @@ namespace DotVVM.Framework.Compilation.Binding
                 return Expression.Call(CreateDelegateMethodInfo, Expression.Constant(delegateType), Target, Expression.Constant(methodInfo))
                     .Apply(e => Expression.Convert(e, delegateType));
         }
-        public Expression CreateMethodCall(IEnumerable<Expression> args)
+        public Expression CreateMethodCall(IEnumerable<Expression> args, MemberExpressionFactory memberExpressionFactory)
         {
             var argsArray = args.ToArray();
             if (IsStatic)
             {
-                return ExpressionHelper.CallMethod((Target as StaticClassIdentifierExpression).Type, BindingFlags.Static | BindingFlags.Public | BindingFlags.FlattenHierarchy, MethodName, TypeArgs, argsArray);
+                return memberExpressionFactory.CallMethod((Target as StaticClassIdentifierExpression).Type, BindingFlags.Static | BindingFlags.Public | BindingFlags.FlattenHierarchy, MethodName, TypeArgs, argsArray);
             }
             else
             {
-                return ExpressionHelper.CallMethod(Target, BindingFlags.Instance | BindingFlags.Public | BindingFlags.FlattenHierarchy, MethodName, TypeArgs, argsArray);
+                return memberExpressionFactory.CallMethod(Target, BindingFlags.Instance | BindingFlags.Public | BindingFlags.FlattenHierarchy, MethodName, TypeArgs, argsArray);
             }
         }
