+++ src/DotVVM.Framework/Compilation/Binding/UnknownTypeSentinel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DotVVM.Framework.Compilation.Binding
+{
+    internal sealed class UnknownTypeSentinel
+    {
+    }
+}
